requirements.txt
pandas>=2.0.3
openai>=1.30.0
python-dotenv>=1.0.0
streamlit>=1.35
tenacity>=8.2.3
tqdm>=4.66.0
openpyxl>=3.1.2
numpy>=2.1.1
